requirements.txt
asgiref==3.11.0
attrs==25.4.0
coverage==7.12.0
Django==5.2.8
django-cors-headers==4.6.0
django-filter==25.2
django-rest-framework==0.1.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
gunicorn==22.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
psycopg2-binary==2.9.9
pyclean==3.4.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==5.0.8
referencing==0.37.0
rpds-py==0.29.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
whitenoise==6.7.0
